@@ -2,7 +2,7 @@
   "name": "hmpps-book-secure-move-frontend",
   "description": "Frontend rendering app for the Book a secure move service",
   "homepage": "https://github.com/ministryofjustice/hmpps-book-secure-move-frontend#readme",
-  "version": "3.0.0",
+  "version": "4.0.0",
   "main": "start.js",
   "contributors": [
     {
@@ -155,7 +155,7 @@
     "accessible-autocomplete": "2.0.4",
     "agentkeepalive": "4.2.1",
     "applicationinsights": "^2.6.0",
-    "axios": "1.7.2",
+    "axios": "1.8.2",
     "axios-debug-log": "1.0.0",
     "bluebird": "3.7.2",
     "chrono-node": "2.5.0",
@@ -175,7 +175,7 @@
     "devour-client": "3.1.5",
     "dotenv": "16.0.3",
     "dropzone": "5.9.3",
-    "express": "4.19.2",
+    "express": "4.21.2",
     "express-session": "1.18.0",
     "faker": "5.5.3",
     "filesize": "10.0.6",
@@ -188,9 +188,11 @@
     "hc-sticky": "2.2.7",
     "helmet": "6.0.1",
     "hmpo-form-wizard": "13.0.0",
+    "hmrc-frontend": "6.60.0",
     "i18next": "22.4.10",
     "i18next-express-middleware": "2.0.0",
     "i18next-fs-backend": "2.1.1",
+    "jquery": "3.7.1",
     "js-yaml": "4.1.0",
     "json2csv": "5.0.7",
     "jsonwebtoken": "9.0.0",
@@ -198,8 +200,10 @@
     "lru-cache": "7.17.0",
     "markdown-it": "13.0.1",
     "markdown-it-container": "3.0.0",
+    "moment": "2.30.1",
     "morgan": "1.10.0",
-    "multer": "^1.4.5-lts.1",
+    "multer": "^2.0.0",
+    "netmask": "2.0.2",
     "node-redis-retry-strategy": "2.1.1",
     "npm-run-all": "4.1.5",
     "nunjucks": "3.2.4",
@@ -223,7 +227,7 @@
     "yargs": "17.7.1"
   },
   "devDependencies": {
-    "@babel/core": "7.23.7",
+    "@babel/core": "7.27.1",
     "@babel/preset-env": "7.20.2",
     "@commitlint/cli": "17.4.4",
     "@commitlint/config-conventional": "17.4.4",
@@ -235,6 +239,7 @@
     "@types/lodash": "4.14.192",
     "@types/mocha": "10.0.1",
     "@types/mock-fs": "4.13.4",
+    "@types/netmask": "2.0.5",
     "@types/node": "18.14.0",
     "@types/nunjucks": "3.2.6",
     "@types/pluralize": "0.0.29",
@@ -310,8 +315,8 @@
     "webpack-merge": "5.8.0"
   },
   "engines": {
-    "node": "20",
-    "npm": "10"
+    "node": "22",
+    "npm": "11"
   },
   "files": [],
   "scripts": {
@@ -371,5 +376,8 @@
   "lint-staged": {
     "*.js": "eslint --cache --fix",
     "package.json": "finepack"
+  },
+  "overrides": {
+    "underscore": "^1.13.6"
   }
 }
