@@ -1,13 +1,15 @@
 const proxyquire = require('proxyquire')
 
-const userSuccessStub = { loadUser: () => Promise.resolve('user') }
+const userSuccessStub = {
+  loadUser: () => Promise.resolve({ id: 'id', permissions: [] }),
+}
 const userFailureError = new Error('test')
 const userFailureStub = { loadUser: () => Promise.reject(userFailureError) }
 
 const expiryTime = 1000
 
 const encodedAccessTokenPayload = Buffer.from(
-  JSON.stringify({ exp: expiryTime })
+  JSON.stringify({ exp: expiryTime, auth_source: 'auth' })
 ).toString('base64')
 
 const accessToken = `test.${encodedAccessTokenPayload}.test`
@@ -141,8 +143,10 @@ describe('Authentication middleware', function () {
             expect(req.session.authExpiry).to.equal(expiryTime)
           })
 
-          it('sets the user info on the session', function () {
-            expect(req.session.user).to.equal('user')
+          it('sets the user info on the session', async function () {
+            expect(req.session.user).to.deep.equal(
+              await userSuccessStub.loadUser()
+            )
           })
 
           it('sets the redirect URL in the session', function () {
