@@ -1,16 +1,17 @@
 const { cloneDeep } = require('lodash')
 
-const commonDateField = require('./common.date')
+const commonDatePickerField = require('./common.date-picker')
 
-const dateCustom = {
-  ...cloneDeep(commonDateField),
-  validate: [...commonDateField.validate, 'required', 'after'],
+const datePicker = {
+  ...cloneDeep(commonDatePickerField),
+  validate: [...commonDatePickerField.validate, 'required', 'after'],
   id: 'date_custom',
   name: 'date_custom',
+  classes: 'govuk-input__wrapper',
   label: {
     text: 'fields::date_custom.label',
     classes: 'govuk-label--s',
   },
 }
 
-module.exports = dateCustom
+module.exports = datePicker
