@@ -112,7 +112,7 @@ describe('Collection controllers', function () {
 
         it('should contain correct number of properties', function () {
           const params = res.render.args[0][1]
-          expect(Object.keys(params)).to.have.length(9)
+          expect(Object.keys(params)).to.have.length(10)
         })
       })
 
@@ -148,7 +148,7 @@ describe('Collection controllers', function () {
 
         it('should contain correct number of properties', function () {
           const params = res.render.args[0][1]
-          expect(Object.keys(params)).to.have.length(10)
+          expect(Object.keys(params)).to.have.length(11)
         })
       })
 
@@ -189,7 +189,7 @@ describe('Collection controllers', function () {
 
         it('should contain correct number of properties', function () {
           const params = res.render.args[0][1]
-          expect(Object.keys(params)).to.have.length(10)
+          expect(Object.keys(params)).to.have.length(11)
         })
       })
     })
