@@ -2,6 +2,7 @@ function switchPeriod(defaults = {}) {
   return function handleSwitchPeriod(req, res) {
     const { view } = req.params
     let currentPeriod = req.session.period
+    const groupBy = req.session.group_by
 
     if (!currentPeriod) {
       currentPeriod = defaults[view]
@@ -12,7 +13,7 @@ function switchPeriod(defaults = {}) {
     const redir = req.originalUrl
       .replace('/switch-view', '')
       .replace(currentPeriod, switchedPeriod)
-    res.redirect(`${redir}`)
+    res.redirect(groupBy ? `${redir}?group_by=${groupBy}` : `${redir}`)
   }
 }
 
