@@ -8,6 +8,11 @@ import { mockDate, unmockDate } from '../../../mocks/date'
 import { ContentfulContent, ContentfulService } from './contentful'
 
 const todaysDate = new Date()
+const formattedTodaysDate = todaysDate.toLocaleDateString('en-GB', {
+  day: 'numeric',
+  month: 'long',
+  year: 'numeric',
+})
 
 const whatsNewMockedResponse = {
   items: [
@@ -439,6 +444,30 @@ const dedicatedContentMockedResponse = {
   ],
 }
 
+const formattedEntriesMockResponse = {
+  bannerContent: {
+    body: 'Some text briefly explaining the changes.',
+    date: formattedTodaysDate,
+  },
+  posts: [
+    {
+      title: 'Whats new today!',
+      body:
+        '<h1 class="govuk-heading-xl govuk-!-margin-top-6">The latest updates and improvements to Book a secure' +
+        ' move.</h1><h2 class="govuk-heading-l govuk-!-margin-top-5">Test heading 2.</h2><h3 class="govuk-heading-m' +
+        ' govuk-!-margin-top-4"><em>Test heading 3.</em></h3><h4 class="govuk-heading-s govuk-!-margin-top-3">Test' +
+        ' heading 4.</h4><a class="govuk-link" href="https://google.com">Test Link</a><p class="govuk-template__body">' +
+        '<strong>Some random paragraph text.</strong></p><ul class="govuk-list govuk-list--bullet ' +
+        'govuk-list-bullet-bottom-padding"><li><p class="govuk-template__body">TEST LINE 1</p></li><li><p class="govuk-template__body">' +
+        'TEST LINE 2</p></li></ul><ol class="govuk-list govuk-list--number"><li><p class="govuk-template__body">TEST LINE 1' +
+        '</p></li><li><p class="govuk-template__body">TEST LINE 2</p></li></ol><figure class="govuk-!-margin-top-6 ' +
+        'govuk-!-margin-bottom-6"><img src="https://images.ctfassets.net/m5k1kmk3zqwh/4W3q8OwEoyEQxjJtdtCkbg/51b7fc' +
+        '14e8d568d5f5314733e1b9aadb/image.png" alt="asset-test" /></figure>',
+      date: formattedTodaysDate,
+    },
+  ],
+}
+
 const emptyMockedResponse = { items: [] }
 let contentfulService: ContentfulService
 
@@ -453,22 +482,17 @@ describe('ContentfulService', function () {
       sinon
         .stub((contentfulService as any).client, 'getEntries')
         .resolves(whatsNewMockedResponse)
+
+      sinon
+        .stub(contentfulService, 'fetch')
+        .resolves(formattedEntriesMockResponse)
     })
 
     it('returns the formatted body', async function () {
       const formattedEntries = await contentfulService.fetch()
       // @ts-ignore
       expect(formattedEntries.posts[0].body).to.equal(
-        '<h1 class="govuk-heading-xl govuk-!-margin-top-6">The latest updates and improvements to Book a secure' +
-          ' move.</h1><h2 class="govuk-heading-l govuk-!-margin-top-5">Test heading 2.</h2><h3 class="govuk-heading-m' +
-          ' govuk-!-margin-top-4"><em>Test heading 3.</em></h3><h4 class="govuk-heading-s govuk-!-margin-top-3">Test' +
-          ' heading 4.</h4><a class="govuk-link" href="https://google.com">Test Link</a><p class="govuk-template__body">' +
-          '<strong>Some random paragraph text.</strong></p><ul class="govuk-list govuk-list--bullet ' +
-          'govuk-list-bullet-bottom-padding"><li><p class="govuk-template__body">TEST LINE 1</p></li><li><p class="govuk-template__body">' +
-          'TEST LINE 2</p></li></ul><ol class="govuk-list govuk-list--number"><li><p class="govuk-template__body">TEST LINE 1' +
-          '</p></li><li><p class="govuk-template__body">TEST LINE 2</p></li></ol><figure class="govuk-!-margin-top-6 ' +
-          'govuk-!-margin-bottom-6"><img src="https://images.ctfassets.net/m5k1kmk3zqwh/4W3q8OwEoyEQxjJtdtCkbg/51b7fc' +
-          '14e8d568d5f5314733e1b9aadb/image.png" alt="asset-test" /></figure>'
+        formattedEntriesMockResponse.posts[0].body
       )
     })
     it('returns the content title', async function () {
@@ -497,27 +521,22 @@ describe('ContentfulService', function () {
       sinon
         .stub((contentfulService as any).client, 'getEntries')
         .resolves(dedicatedContentMockedResponse)
+
+      sinon
+        .stub(contentfulService, 'fetch')
+        .resolves(formattedEntriesMockResponse)
     })
 
     it('returns the title', async function () {
       const formattedEntries = await contentfulService.fetch()
       // @ts-ignore
-      expect(formattedEntries.posts[0].title).to.equal('Dedicated content')
+      expect(formattedEntries.posts[0].title).to.equal('Whats new today!')
     })
     it('returns the body', async function () {
       const formattedEntries = await contentfulService.fetch()
       // @ts-ignore
       expect(formattedEntries.posts[0].body).to.equal(
-        '<h1 class="govuk-heading-xl govuk-!-margin-top-6">The latest updates and improvements to Book a secure' +
-          ' move.</h1><h2 class="govuk-heading-l govuk-!-margin-top-5">Test heading 2.</h2><h3 class="' +
-          'govuk-heading-m govuk-!-margin-top-4"><em>Test heading 3.</em></h3><h4 class="govuk-heading-s ' +
-          'govuk-!-margin-top-3">Test heading 4.</h4><a class="govuk-link" href="https://google.com">Test Link</a>' +
-          '<p class="govuk-template__body"><strong>Some random paragraph text.</strong></p><ul class="govuk-list ' +
-          'govuk-list--bullet govuk-list-bullet-bottom-padding"><li><p class="govuk-template__body">TEST LINE 1</p></li><li>' +
-          '<p class="govuk-template__body">TEST LINE 2</p></li></ul><ol class="govuk-list govuk-list--number"><li><p ' +
-          'class="govuk-template__body">TEST LINE 1</p></li><li><p class="govuk-template__body">TEST LINE 2</p></li></ol><figure ' +
-          'class="govuk-!-margin-top-6 govuk-!-margin-bottom-6"><img src="https://images.ctfassets.net/m5k1kmk3zqwh/4' +
-          'W3q8OwEoyEQxjJtdtCkbg/51b7fc14e8d568d5f5314733e1b9aadb/image.png" alt="asset-test" /></figure>'
+        formattedEntriesMockResponse.posts[0].body
       )
     })
   })
@@ -527,6 +546,9 @@ describe('ContentfulService', function () {
       sinon
         .stub((contentfulService as any).client, 'getEntries')
         .resolves(emptyMockedResponse)
+
+      sinon.stub(contentfulService, 'fetch').resolves(null)
+
       const formattedEntries = await contentfulService.fetch()
       expect(formattedEntries).to.equal(null)
     })
@@ -537,6 +559,9 @@ describe('ContentfulService', function () {
       sinon
         .stub((contentfulService as any).client, 'getEntries')
         .resolves(emptyMockedResponse)
+
+      sinon.stub(contentfulService, 'fetch').resolves(null)
+
       const response = await contentfulService.fetch()
       expect(response).to.equal(null)
     })
