@@ -1,19 +1,36 @@
 import I18n from '../../config/i18n'
+import { EventService } from '../services/event'
 import { LodgingService } from '../services/lodging'
+import { SupplierService } from '../services/supplier'
 
 import { Allocation } from './allocation'
+import { BasmError } from './basm_error'
 import { Journey } from './journey'
 import { Lodging } from './lodging'
 import { Move } from './move'
 import { Service } from './service'
 import { SessionModel } from './session_model'
-import { SupplierService } from "../services/supplier"
-import { EventService } from '../services/event'
+
+interface Session {
+  authExpiry: number
+  save: () => void
+  regenerate: (func: (error: BasmError) => void) => void
+  user: {
+    id: string
+    permissions: never[]
+  }
+  [key: string]: any
+}
 
 export interface BasmRequest extends Express.Request {
-  body: any
   allocation?: Allocation
+  body: any
   canAccess: (permission: string) => boolean
+  connection: {
+    remoteAddress?: string
+    socket?: { remoteAddress?: string }
+  }
+  socket: { remoteAddress?: string }
   flash: (yeah: any, nah: any) => void
   form?: {
     options: {
@@ -24,15 +41,14 @@ export interface BasmRequest extends Express.Request {
     }
     values: { [key: string]: any }
   }
+  headers: { [key: string]: string }
   initialStep?: any
   journeyModel: SessionModel
   journeys: Journey[]
   move: Move
   lodging?: Lodging
   params: any
-  session: {
-    save: () => void
-  }
+  session: Session
   sessionModel: SessionModel
   services: {
     allocation: Service
@@ -41,8 +57,10 @@ export interface BasmRequest extends Express.Request {
     event: EventService
   }
   t: typeof I18n.t
-  user: { 
-    id: string,
-    permissions: never[],
-  },
+  user: {
+    id: string
+    permissions: never[]
+    username: string
+  }
+  url: string
 }
