@@ -16,9 +16,9 @@ module.exports = {
   mojFrontend: path.resolve(
     root,
     'node_modules',
-    '@ministryofjustice',
-    'frontend'
+    '@ministryofjustice/frontend'
   ),
+  hmrcFrontend: path.resolve(root, 'node_modules', 'hmrc-frontend'),
   fixtures: path.resolve(root, 'test', 'fixtures'),
   frameworks: {
     source: path.resolve(
