@@ -1,12 +1,13 @@
 import { Factory } from 'fishery'
 
+import { EventService } from '../common/services/event'
 import { LodgingService } from '../common/services/lodging'
+import { SupplierService } from '../common/services/supplier'
+import { BasmError } from '../common/types/basm_error'
 import { BasmRequest } from '../common/types/basm_request'
 import I18n from '../config/i18n'
 
 import { MoveFactory } from './move'
-import {SupplierService} from "../common/services/supplier"
-import { EventService } from '../common/services/event'
 
 export const defaultParams = {
   canAccess: (permission: string) => true,
@@ -19,7 +20,13 @@ export const defaultParams = {
   move: MoveFactory.build(),
   params: {},
   session: {
+    authExpiry: 17416129189670,
+    regenerate: (_func: (error: BasmError) => void) => {},
     save: () => {},
+    user: {
+      id: '',
+      permissions: [],
+    },
   },
   sessionModel: {
     reset: () => {},
@@ -40,7 +47,12 @@ export const defaultParams = {
   user: {
     id: '',
     permissions: [],
+    username: '',
   },
+  headers: {},
+  connection: {},
+  socket: {},
+  url: '',
 }
 
 export const BasmRequestFactory = Factory.define<BasmRequest>(() => ({
