@@ -1,9 +1,6 @@
 import { addDays } from 'date-fns'
-import { fixture } from 'testcafe'
+import { Selector } from 'testcafe'
 
-import { createMove } from '../_move'
-import { pmuUser } from '../_roles'
-import { home } from '../_routes'
 import {
   lodgingCancelPage,
   moveDetailPage,
@@ -11,14 +8,9 @@ import {
   moveLodgeLocationPage,
   moveLodgeSavedPage,
   page,
-} from '../pages'
+} from '../../pages'
 
-fixture('A move without any lodges').beforeEach(async t => {
-  await t.useRole(pmuUser).navigateTo(home)
-  await createMove()
-})
-
-async function testLodges(t: any, count: number, length: number) {
+export async function testLodges(t: any, count: number, length: number) {
   const lodgingLocations: string[] = []
 
   for (let i = 0; i < count; i++) {
@@ -89,27 +81,3 @@ async function testLodges(t: any, count: number, length: number) {
   // Check lodge was cancelled
   await moveDetailPage.checkLodgesInDetails(0)
 }
-
-test('Without any existing lodges', async function (t) {
-  await testLodges(t, 1, 1)
-})
-
-test('Without any existing lodges - two days', async function (t) {
-  await testLodges(t, 1, 2)
-})
-
-test('Without any existing lodges - seven days', async function (t) {
-  await testLodges(t, 1, 7)
-})
-
-test('Without any existing lodges - two lodges', async function (t) {
-  await testLodges(t, 2, 1)
-})
-
-test('Without any existing lodges - two lodges - two days', async function (t) {
-  await testLodges(t, 2, 2)
-})
-
-test('Without any existing lodges - two lodges - seven days', async function (t) {
-  await testLodges(t, 2, 7)
-})
